requirements.txt
streamlit
pandas
openpyxl
python-dotenv
psycopg2-binary
numpy